[xds_weighted_target] balancergroup test passed
